@@ -1,40 +1,35 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Vejrudsigten.Services;
 
-namespace Vejrudsigten.Pages
+namespace Vejrudsigten.Pages;
+
+public class IndexModel : PageModel
 {
-    public class IndexModel : PageModel
+    private readonly IConfiguration _configuration;
+    private readonly ILogger<IndexModel> _logger;
+
+    public IndexModel(ILogger<IndexModel> logger, IConfiguration configuration)
     {
-        private readonly ILogger<IndexModel> _logger;
-        private readonly IConfiguration _configuration;
+        _logger = logger;
+        _configuration = configuration;
+    }
 
-        public IndexModel(ILogger<IndexModel> logger, IConfiguration configuration)
+    public async Task OnGetAsync()
+    {
+        var key = _configuration["key"];
+
+        if (key == null)
         {
-            _logger = logger;
-            _configuration = configuration;
+            ViewData.Add("Vejrudsigten",
+                "Hov! Du har glemt at angive nøglen i appsettings.local.json. Gå tilbage til opgavebeskrivelsen og se hvordan");
         }
-
-        public async Task OnGetAsync()
+        else
         {
-            var key = _configuration["key"];
-
-            if (key == null)
-            {
-                ViewData.Add("Vejrudsigten", "Hov! Du har glemt at angive nøglen i appsettings.local.json. Gå tilbage til opgavebeskrivelsen og se hvordan");
-            } else
-            {
-                ViewData.Add("Vejrudsigten", await WeatherForecast.GetForecastAsync(key));
-            }
-
-            
-
+            var weatherService = new WeatherService();
+            ViewData.Add("Vejrudsigten", await new WeatherForecast(weatherService).GetForecastAsync(key, "Kolding"));
         }
     }
 }
